Updated Component Ribbon
