+++ Plugin/Deprecated/200908_SetDO.cs
@@ -15,7 +15,7 @@ namespace Axis.IO
         // Context menu data items
         bool m_Sync = false;
 
-        public SetDO_Obsolete() : base("Set Digital Output", "Set DO", "Set the value of a digital output.", AxisInfo.Plugin, AxisInfo.TabIO)
+        public SetDO_Obsolete() : base("Set Digital Output", "Set DO", "Set the value of a digital output.", AxisInfo.Plugin, AxisInfo.TabDepricated)
         {
         }
