+++ Plugin/Targets/GH_CreateTarget_Multithreaded.cs
@@ -2,6 +2,10 @@
 using System.Collections.Generic;
 using System.Linq;
 using System.Windows.Forms;
+using System.Threading.Tasks;
+
+//using MoreLinq;
+
 using Grasshopper.Kernel;
 using Grasshopper.Kernel.Parameters;
 using Grasshopper.Kernel.Types;
@@ -14,22 +18,21 @@ namespace Axis
     /// <summary>
     /// Define robot end effector plane target positions.
     /// </summary>
-    public class CreateTarget : GH_Component, IGH_VariableParameterComponent
+    public class GH_CreateTarget_Multithreaded : GH_Component, IGH_VariableParameterComponent
     {
         // Boolean toggle for context menu items.
         bool m_outputCode = false;
         bool m_interpolationTypes = false;
-        bool m_outputTarget = false;
 
         Manufacturer m_Manufacturer = Manufacturer.ABB;
-        List<Target> m_targets = new List<Target>();
-        BoundingBox m_bBox = new BoundingBox();
+        List<Target> c_targets = new List<Target>();
+        BoundingBox c_bBox = new BoundingBox();
 
         // External axis presence.
         bool extRotary = false;
         bool extLinear = false;
-        
-        public CreateTarget() : base("Plane Target", "Target", "Create custom robot targets from planes.", AxisInfo.Plugin, AxisInfo.TabTargets)
+
+        public GH_CreateTarget_Multithreaded() : base("Plane Target", "Target", "Create custom robot targets from planes.", AxisInfo.Plugin, AxisInfo.TabConfiguration)
         {
         }
 
@@ -46,14 +49,14 @@ namespace Axis
             IGH_Param csystem = new Axis.Params.CSystemParam();
             pManager.AddParameter(csystem, "Wobj", "Wobj", "Wobj to use for operation.", GH_ParamAccess.list);
 
-            for (int i = 0; i < 5; i++)
+            for (int i = 1; i < 5; i++)
                 pManager[i].Optional = true;
         }
 
         protected override void RegisterOutputParams(GH_Component.GH_OutputParamManager pManager)
         {
             IGH_Param target = new Axis.Params.TargetParam();
-            pManager.AddParameter(target,"Targets", "Targets", "Robot targets.", GH_ParamAccess.list);
+            pManager.AddParameter(target, "Targets", "Targets", "Robot targets.", GH_ParamAccess.list);
         }
         #endregion
 
@@ -67,9 +70,11 @@ namespace Axis
         {
             this.Message = m_Manufacturer.ToString();
 
+            List<string> code = new List<string>();
+
             List<Plane> planes = new List<Plane>();
-            List<GH_ObjectWrapper> speedsIn = new List<GH_ObjectWrapper>();
-            List<GH_ObjectWrapper> zonesIn = new List<GH_ObjectWrapper>();
+            List<Speed> speeds = new List<Speed>();
+            List<Zone> zones = new List<Zone>();
             List<Tool> tools = new List<Tool>();
             List<CSystem> wobjs = new List<CSystem>();
             List<int> methods = new List<int>();
@@ -78,12 +83,11 @@ namespace Axis
             List<double> eRotVals = new List<double>();
             List<double> eLinVals = new List<double>();
 
-            bool hasSpeed = true;
-            bool hasZone = true;
 
-            if (!DA.GetDataList(0, planes)) return;
-            if (!DA.GetDataList(1, speedsIn)) hasSpeed = false;
-            if (!DA.GetDataList(2, zonesIn)) hasZone = false;
+
+            DA.GetDataList(0, planes);
+            if (!DA.GetDataList(1, speeds)) speeds.Add(Speed.Default);
+            if (!DA.GetDataList(2, zones)) zones.Add(Zone.Default);
             if (!DA.GetDataList(3, tools)) tools.Add(Tool.Default);
             if (!DA.GetDataList(4, wobjs)) wobjs.Add(CSystem.Default);
 
@@ -94,151 +98,124 @@ namespace Axis
             if (extRotary) { if (!DA.GetDataList("Rotary", eRotVals)) return; }
             if (extLinear) { if (!DA.GetDataList("Linear", eLinVals)) return; }
 
-            List<Speed> speeds = new List<Speed>();
-            List<Zone> zones = new List<Zone>();
 
-            // Check to see if we have speeds, and if they are custom speed objects, otherwise use values.
-            if (hasSpeed)
-            {
-                // Default speed dictionary.
-                Dictionary<double, Speed> defaultSpeeds = Util.ABBSpeeds();
-                double speedVal = 0;
-
-                foreach (GH_ObjectWrapper speedIn in speedsIn)
-                {
-                    GH_ObjectWrapper speedObj = speedIn;
-                    Type cType = speedObj.Value.GetType();
-                    GH_Convert.ToDouble_Secondary(speedObj.Value, ref speedVal);
-
-                    if (cType.Name == "Speed")
-                        speeds.Add(speedObj.Value as Speed);
-                    else
-                    {
-                        if (!defaultSpeeds.ContainsKey(speedVal))
-                            AddRuntimeMessage(GH_RuntimeMessageLevel.Warning, "Supplied speed value is non-standard. Please supply a default value (check the Axis Wiki - Controlling Speed for more info) or create a custom speed using the Speed component.");
-                        else
-                            speeds.Add(defaultSpeeds[speedVal]);
-                    }
-                }
-            }
-            // If we don't have any speed values, use the default speed.
-            else
-                speeds.Add(Speed.Default);
-
-            // Check to see if we have zones, and if they are custom zones objects, otherwise use values.
-            if (hasZone)
-            {
-                // Default zone dictionary.
-                Dictionary<double, Zone> defaultZones = Util.ABBZones();
-                double zoneVal = 0;
-
-                foreach (GH_ObjectWrapper zoneIn in zonesIn)
-                {
-                    GH_ObjectWrapper zoneObj = zoneIn;
-                    Type cType = zoneObj.Value.GetType();
-                    GH_Convert.ToDouble_Secondary(zoneObj.Value, ref zoneVal);
-
-                    if (cType.Name == "Zone")
-                        zones.Add(zoneObj.Value as Zone);
-                    else
-                    {
-                        if (!defaultZones.ContainsKey(zoneVal))
-                            AddRuntimeMessage(GH_RuntimeMessageLevel.Warning, "Supplied zone value is non-standard. Please supply a default value (check the Axis Wiki - Controlling Zone for more info) or create a custom zone using the Zoe component.");
-                        else
-                            zones.Add(defaultZones[zoneVal]);
-                    }
-                }
-            }
-            // If we don't have any zone values, use the default zone.
-            else
-                zones.Add(Zone.Default);
-
-            List<Target> targets = new List<Target>();
-            List<string> code = new List<string>();
-            Speed speed = Speed.Default;
-            Zone zone = Zone.Default;
-            Tool tool = Tool.Default;
-            CSystem wobj = CSystem.Default;
-            int method = 0;
+            // Compute results on given data
+            System.Collections.Concurrent.ConcurrentBag<TargetOrderPair> bag = new System.Collections.Concurrent.ConcurrentBag<TargetOrderPair>();
+            //Parallel.For(0, planes.Count, index => ComputeTargets_Unordered(bag, index, planes, m_interpolationTypes, speeds, zones, extRotary, extLinear, tools, wobjs, eRotVals, eLinVals, methods, m_Manufacturer));
+            Parallel.For(0, planes.Count, index => ComputeTargets(bag, index, planes, speeds, zones, tools, wobjs, eRotVals, eLinVals, methods, m_Manufacturer));
+            c_targets = bag.OrderBy(b => b.Order).Select(b => b.Value).ToList();
             
-            // External axis placeholders
-            double extRot = Util.ExAxisTol;
-            double extLin = Util.ExAxisTol;
+            //Compute bounding box for visualisation
+            c_bBox = new BoundingBox(c_targets.Select(t => t.Plane.Origin).ToList());
 
-            for (int i = 0; i < planes.Count; i++)
+            // Set output data
+            if (c_targets != null)
             {
-                if (m_interpolationTypes)
-                {
-                    // Method
-                    if (i < methods.Count) { method = methods[i]; }
-                    else if (methods != null && i >= methods.Count) { method = methods[methods.Count - 1]; }
-                    else { method = 0; }
-                }
-
-                if (speeds.Count > 0)
-                {
-                    if (i < speeds.Count) { speed = speeds[i]; }
-                    else { speed = speeds[speeds.Count - 1]; }
-                }
-                else { speed = Speed.Default; }
-
-                // Zone
-                if (i < zones.Count) { zone = zones[i]; }
-                else { zone = zones[zones.Count - 1]; }
-
-                // External rotary axis
-                if (extRotary)
-                {
-                    if (i < eRotVals.Count) { extRot = Math.Round(eRotVals[i], 3); }
-                    else { extRot = Math.Round(eRotVals[eRotVals.Count - 1], 3); }
-                }
-
-                // External linear axis
-                if (extLinear)
-                {
-                    if (i < eLinVals.Count) { extLin = Math.Round(eLinVals[i], 3); }
-                    else { extLin = Math.Round(eLinVals[eLinVals.Count - 1], 3); }
-                }
-
-                // Tools
-                if (tools.Count > 0)
-                {
-                    if (i < tools.Count) { tool = tools[i]; }
-                    else { tool = tools[tools.Count - 1]; }
-                }
-                else { tool = Tool.Default; }
-
-                // Wobjs
-                if (wobjs.Count > 0)
-                {
-                    if (i < wobjs.Count) { wobj = wobjs[i]; }
-                    else { wobj = wobjs[wobjs.Count - 1]; }
-                }
-                else { wobj = CSystem.Default; }
-
-                // Methods
-                MotionType mType = MotionType.Linear;
-
-                if (method == 1) { mType = MotionType.Joint; }
-                else if (method == 2) { mType = MotionType.AbsoluteJoint; }
-
-                // Create the robot target.
-                Target robTarg = new Target(planes[i], mType, speed, zone, tool, wobj, extRot, extLin, m_Manufacturer);
-                targets.Add(robTarg);
-
-                code.Add(robTarg.StrRob);
+                DA.SetDataList("Targets", c_targets);
             }
-            DA.SetDataList(0, targets);
-
-            m_targets = targets;
-
-            List<Point3d> points = new List<Point3d>();
-            foreach (Target t in targets) points.Add(t.Position);
-            m_bBox = new BoundingBox(points);
-            /*
-            if (m_outputTarget)
+            if (m_outputCode) 
+            {
+                code = c_targets.Select(t => t.StrRob).ToList();
                 DA.SetDataList("Code", code);
-            */
+            }
+        }
+
+        struct TargetOrderPair
+        {
+            public Target Value { get; set; }
+            public int Order { get; set; }
+        }
+
+        private static void ComputeTargets_Unordered(System.Collections.Concurrent.ConcurrentBag<TargetOrderPair> bag, 
+            int i, List<Plane> planes, bool m_interpolationTypes, List<Speed> speeds,
+            List<Zone> zones, bool extRotary, bool extLinear, List<Tool> tools, List<CSystem> wobjs, List<double> eRotVals,
+            List<double> eLinVals, List<int> methods, Manufacturer m_Manufacturer)
+        {
+            TargetOrderPair result = new TargetOrderPair();
+
+            Speed speed;
+            Zone zone;
+            Tool tool;
+            CSystem wobj;
+            int method = 0;
+            double extRot = 0;
+            double extLin = 0;
+
+
+            if (m_interpolationTypes)
+            {
+                // Method
+                if (i < methods.Count) { method = methods[i]; }
+                else if (methods != null && i >= methods.Count) { method = methods[methods.Count - 1]; }
+                else { method = 0; }
+            }
+
+            if (speeds.Count > 0)
+            {
+                if (i < speeds.Count) { speed = speeds[i]; }
+                else { speed = speeds[speeds.Count - 1]; }
+            }
+            else { speed = Speed.Default; }
+
+            // Zone
+            if (i < zones.Count) { zone = zones[i]; }
+            else { zone = zones[zones.Count - 1]; }
+
+            // External rotary axis
+            if (extRotary)
+            {
+                if (i < eRotVals.Count) { extRot = Math.Round(eRotVals[i], 3); }
+                else { extRot = Math.Round(eRotVals[eRotVals.Count - 1], 3); }
+            }
+
+            // External linear axis
+            if (extLinear)
+            {
+                if (i < eLinVals.Count) { extLin = Math.Round(eLinVals[i], 3); }
+                else { extLin = Math.Round(eLinVals[eLinVals.Count - 1], 3); }
+            }
+
+            // Tools
+            if (tools.Count > 0)
+            {
+                if (i < tools.Count) { tool = tools[i]; }
+                else { tool = tools[tools.Count - 1]; }
+            }
+            else { tool = Tool.Default; }
+
+            // Wobjs
+            if (wobjs.Count > 0)
+            {
+                if (i < wobjs.Count) { wobj = wobjs[i]; }
+                else { wobj = wobjs[wobjs.Count - 1]; }
+            }
+            else { wobj = CSystem.Default; }
+
+            // Methods
+            MotionType mType = MotionType.Linear;
+
+            if (method == 1) { mType = MotionType.Joint; }
+            else if (method == 2) { mType = MotionType.AbsoluteJoint; }
+
+            // Create the robot target.
+            result.Value = new Target(planes[i], mType, speed, zone, tool, wobj, extRot, extLin, m_Manufacturer);
+            result.Order = i;
+
+            bag.Add(result);
+        }
+
+        private static void ComputeTargets(System.Collections.Concurrent.ConcurrentBag<TargetOrderPair> bag, int i, 
+            List<Plane> planes, List<Speed> speeds, List<Zone> zones, List<Tool> tools, List<CSystem> wobjs, 
+            List<double> eRotVals, List<double> eLinVals, List<int> methods, Manufacturer m_Manufacturer)
+        {
+            TargetOrderPair result = new TargetOrderPair();
+
+            // Create the robot target.
+            result.Value = new Target(planes[i], (MotionType)methods.InfinitElementAt(i), speeds.InfinitElementAt(i), zones.InfinitElementAt(i), 
+                tools.InfinitElementAt(i), wobjs.InfinitElementAt(i), eRotVals.InfinitElementAt(i), eLinVals.InfinitElementAt(i), m_Manufacturer);
+            result.Order = i;
+
+            bag.Add(result);
         }
 
         #region UI
@@ -275,9 +252,9 @@ namespace Axis
         // Build a list of optional input parameters
         IGH_Param[] inputParams = new IGH_Param[3]
         {
-        new Param_Integer() { Name = "*Method", NickName = "Method", Description = "A list of target interpolation types [0 = Linear, 1 = Joint]. If one value is supplied it will be applied to all targets.", Access = GH_ParamAccess.list },
-        new Param_Number() { Name = "Rotary", NickName = "Rotary", Description = "A list of external rotary axis positions in degrees. If one value is supplied it will be applied to all targets.", Access = GH_ParamAccess.list },
-        new Param_Number() { Name = "Linear", NickName = "Linear", Description = "A list of external linear axis positions in degrees. If one value is supplied it will be applied to all targets.", Access = GH_ParamAccess.list },
+            new Param_Integer() { Name = "*Method", NickName = "Method", Description = "A list of target interpolation types [0 = Linear, 1 = Joint]. If one value is supplied it will be applied to all targets.", Access = GH_ParamAccess.list },
+            new Param_Number() { Name = "Rotary", NickName = "Rotary", Description = "A list of external rotary axis positions in degrees. If one value is supplied it will be applied to all targets.", Access = GH_ParamAccess.list },
+            new Param_Number() { Name = "Linear", NickName = "Linear", Description = "A list of external linear axis positions in degrees. If one value is supplied it will be applied to all targets.", Access = GH_ParamAccess.list },
         };
 
         // Build a list of optional output parameters
@@ -289,7 +266,7 @@ namespace Axis
         // The following functions append menu items and then handle the item clicked event.
         protected override void AppendAdditionalComponentMenuItems(System.Windows.Forms.ToolStripDropDown menu)
         {
-            ToolStripMenuItem outputCode = Menu_AppendItem(menu, "Output Code", outputCode_Click, true, m_outputTarget);
+            ToolStripMenuItem outputCode = Menu_AppendItem(menu, "Output Code", outputCode_Click, true, m_outputCode);
             outputCode.ToolTipText = "Output a string representation of the robot targets.";
 
             ToolStripSeparator seperator = Menu_AppendSeparator(menu);
@@ -328,9 +305,9 @@ namespace Axis
         private void outputCode_Click(object sender, EventArgs e)
         {
             RecordUndoEvent("OutputCode");
-            m_outputTarget = !m_outputTarget;
+            m_outputCode = !m_outputCode;
 
-            if (m_outputTarget)
+            if (m_outputCode)
             {
                 AddOutput(0);
             }
@@ -447,21 +424,21 @@ namespace Axis
         public override void DrawViewportMeshes(IGH_PreviewArgs args)
         {
             base.DrawViewportMeshes(args);
-            foreach (Target target in m_targets) Canvas.Component.DisplayPlane(target.Plane, args);
+            foreach (Target target in c_targets) Canvas.Component.DisplayPlane(target.Plane, args);
         }
         public override void DrawViewportWires(IGH_PreviewArgs args)
         {
 
             base.DrawViewportWires(args);
-            foreach (Target target in m_targets) Canvas.Component.DisplayPlane(target.Plane, args);
+            foreach (Target target in c_targets) Canvas.Component.DisplayPlane(target.Plane, args);
 
         }
-        
+
         public override void ClearData()
         {
             base.ClearData();
-            m_targets.Clear();
-            m_bBox = BoundingBox.Empty;
+            c_targets.Clear();
+            c_bBox = BoundingBox.Empty;
         }
         #endregion
 
@@ -509,8 +486,18 @@ namespace Axis
         }
         public override Guid ComponentGuid
         {
-            get { return new Guid("{c8ae5262-f447-4807-b1ff-10b29b37c984}"); }
+            get { return new Guid("C3B49429-C457-4573-939E-0D3BA69DF071"); }
         }
         #endregion
+    }
+
+    public static class Extentions
+    {
+        public static T InfinitElementAt<T>(this IList<T> source, int index) 
+        {
+            if (source == null) return default(T);
+            if (source.Count() == 0) return default(T);
+            return ((source.Count() > index) ? source[index] : source[source.Count()-1]); 
+        }
     }
 }
