+++ Plugin/Targets/GH_FlipPlane.cs
@@ -9,9 +9,9 @@ namespace Axis.Targets
     /// <summary>
     /// Flip a plane to comply with robot programming conventions.
     /// </summary>
-    public class FlipPlane : GH_Component
+    public class GH_FlipPlane : GH_Component
     {
-        public FlipPlane() : base("Flip Plane", "Flip", "Flip a plane around it's Y axis.", AxisInfo.Plugin, AxisInfo.TabTargets)
+        public GH_FlipPlane() : base("Flip Plane", "Flip", "Flip a plane around it's Y axis.", AxisInfo.Plugin, AxisInfo.TabConfiguration)
         {
         }
 
@@ -55,7 +55,7 @@ namespace Axis.Targets
         }
 
         #region Component Settings
-        public override GH_Exposure Exposure => GH_Exposure.secondary;
+        public override GH_Exposure Exposure => GH_Exposure.tertiary;
         protected override System.Drawing.Bitmap Icon
         {
             get
